+++ apps/hubble/src/defaultConfig.ts
@@ -10,7 +10,7 @@ const DEFAULT_GOSSIP_PORT = 2282;
 const DEFAULT_RPC_PORT = 2283;
 const DEFAULT_HTTP_API_PORT = 2281;
 const DEFAULT_NETWORK = 3; // Farcaster Devnet
-export const DEFAULT_CATCHUP_SYNC_SNAPSHOT_MESSAGE_LIMIT = 3_000_000;
+export const DEFAULT_CATCHUP_SYNC_SNAPSHOT_MESSAGE_LIMIT = 10_000_000;
 
 export const Config = {
   /** Path to a PeerId file */
@@ -59,7 +59,7 @@ export const Config = {
   /** Catchup sync with snapshot
    * NOTE: Catchup sync using snapshot WILL RESET THE DATABASE
    */
-  catchupSyncWithSnapshot: false,
+  catchupSyncWithSnapshot: true,
   /**
    * Message limit - when exceeded, trigger catchup sync using snapshot
    * NOTE: Catchup sync using snapshot WILL RESET THE DATABASE