+++ packages/hub-shuttle/src/shuttle/hubEventProcessor.ts
@@ -1,7 +1,8 @@
 import { HubEvent, isMergeMessageHubEvent, MergeMessageHubEvent, Message } from "@farcaster/hub-nodejs";
 import { DB } from "./db";
 import { MessageProcessor } from "./messageProcessor";
-import { MessageHandler } from "./interfaces";
+import { MessageHandler } from "./";
+import { log } from "../log";
 
 export class HubEventProcessor {
   static async processHubEvent(db: DB, event: HubEvent, handler: MessageHandler) {
@@ -17,8 +18,8 @@ export class HubEventProcessor {
 
   private static async processMergeMessage(db: DB, message: Message, handler: MessageHandler, wasMissed = false) {
     await db.transaction().execute(async (trx) => {
-      await MessageProcessor.storeMessage(message, trx, "merge");
-      await handler.handleMessageMerge(message, trx, "merge", wasMissed);
+      const isNew = await MessageProcessor.storeMessage(message, trx, "merge", log);
+      await handler.handleMessageMerge(message, trx, "merge", isNew, wasMissed);
     });
   }
 }