+++ packages/hub-shuttle/src/shuttle/index.ts
@@ -1,7 +1,20 @@
 import { Message } from "@farcaster/hub-nodejs";
 import { DB } from "./db";
 
+export * from "./db";
+export * from "./redis";
+export * from "./hub";
+export * from "./hubSubscriber";
+export * from "./hubEventProcessor";
+export * from "./messageProcessor";
+export * from "./messageReconciliation";
+export * from "./eventStream";
+
 export type StoreMessageOperation = "merge" | "delete" | "revoke" | "prune";
+
+export type ProcessResult = {
+  skipped: boolean;
+};
 
 // Implement this interface in your app to handle messages. The package currently provides the following gurantees:
 // - Messages can be assumed to be processes in the same order as they were recieved by the hub as long as wasMissed in false
@@ -10,5 +23,11 @@ export type StoreMessageOperation = "merge" | "delete" | "revoke" | "prune";
 // - The package will not process the same message twice
 
 export interface MessageHandler {
-  handleMessageMerge(message: Message, txn: DB, operation: StoreMessageOperation, wasMissed: boolean): Promise<void>;
+  handleMessageMerge(
+    message: Message,
+    txn: DB,
+    operation: StoreMessageOperation,
+    isNew: boolean,
+    wasMissed: boolean,
+  ): Promise<void>;
 }