+++ packages/hub-shuttle/src/statsd.ts
@@ -1,0 +1,3 @@
+import StatsD from "@figma/hot-shots";
+
+export const statsd = new StatsD();