+++ packages/hub-shuttle/tsup.config.ts
@@ -2,11 +2,7 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   target: ["node14", "chrome79", "edge109", "firefox102", "safari12"],
-  entryPoints: ["./src/**/*.ts"],
-  esbuildOptions(options, context) {
-    options.outbase = "./";
-  },
-  bundle: false,
+  entryPoints: ["./src/index.ts"],
   format: ["esm", "cjs"],
   dts: true,
   clean: true,