+++ packages/hub-shuttle/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @farcaster/hub-shuttle
 
+## 0.1.5
+
+### Patch Changes
+
+- fix: Fix module import issue
+
+## 0.1.4
+
+### Patch Changes
+
+- fix: Fix interfaces import
+
 ## 0.1.3
 
 ### Patch Changes